changed name of main_activity to login_activity created new main activity_main Set up storage with firebase... no code yet, but it is ready to be used
