+++ app/src/main/java/com/bcit/walksforwalks/RegisterUser.java
@@ -39,9 +39,6 @@ public class RegisterUser extends AppCompatActivity implements View.OnClickListe
         progressBar = findViewById(R.id.progressBar_register_user);
 
         registerUser.setOnClickListener(this);
-
-
-
     }
 
     @Override
